Project has been successfully split
Resolved #138
