--- src/main/java/bot/dogbot/README.md
@@ -1,55 +1,0 @@
-Dogbot
-======
-
-This is an implementation of [Chatbot](https://github.com/hollandjake/Chatbot)
-which has been created for the use in the Computer Science Chat at The University Of Sheffield.
-
-Commands
-========
-
-Inherited commands
-------------------
-| Command | Response |
-| ------- | -------- |
-| `!commands` or `!help` | Links to this readme |
-| `!github` or `!repo` | Links to [this](https://github.com/hollandjake/Chatbot) github repository |
-| `!ping` | Checks if bot is active |
-| `!shutdown [code]` | shuts the bot down. The code is output at boot for the bot |
-| `!stats` | Outputs information about the bots configuration |
-| `!uptime` or `!puptime` | Outputs how long the bot has been running |
-
-Image responses
----------------
-| Command | Response |
-| ------- | -------- |
-| `!bird` or `!birb` | Sends a picture of a bird |
-| `!cat` | Sends a picture of a cat |
-| `!dog` | Sends a picture of a dog |
-| `!extragooddog` | Send a picture of a dog from a curated list |
-| `!inspire` | Sends an inspirational quote from [InspiroBot](http://inspirobot.me) |
-| `!react` or `!reac` or `!reac[as many c's]`| Gives the cats reaction |
-| `!tab` | Summons the tabulance to make things better |
-| `!xkcd` | Sends a random XKCD |
-| `!xkcd l` or `!xkcd latest` | Sends the latest XKCD |
-| `!xkcd [num]` | Sends a specific XKCD |
-
-Message responses
------------------
-| Command | Response |
-| ------- | -------- |
-| `!ask [message]` or `!8ball [message]` | Query the all knowing magic 8 ball |
-| `!feedback` | Sends a link to the group [Feedback form](https://docs.google.com/document/d/19Vquu0fh8LCqUXH0wwpm9H9MSq1LrEx1Z2Xg9NknKmg/edit?usp=sharing) |
-| `!grab` | Grabs the previous message |
-| `!grab [num]` | Grabs the message [num] behind. `!grab 1` is equal to `!grab` |
-| `!locate [message]` or `!grab [message]` | Grabs the most recent message to match the query in message |
-| `!quote` | Returns a random grabbed message |
-| `!QUOTE` | Returns a random grabbed message but in uppercase |
-| `!QuOtE` | Returns a random grabbed message but in shakey case |
-| `!quotecount` | Returns number of quotes for the current thread |
-| `!quotecount [name]` | Returns number of quotes from a user in this thread |
-| `!reddits` | Outputs list of reddits being used |
-| `!roll` | Returns a random number between 1-6 |
-| `!roll [num]` | Returns a random number between 1-[num] |
-| `!think` or `!thonk` | Returns a think emoji|
-| `!think [num]` or `!thonk [num]` | Returns [num] amount of think emoji's (limited to 100) |
-| `!trello` | Links to the group [Trello page](https://trello.com/b/9f49WSW0/second-year-compsci) |
