--- src/test/java/Chatbot/Birds/responses.txt
@@ -1,6 +1,0 @@
-Ruffling feathers...
-Tweet twoot
-Quack.
-SQUAWK!
-Do I get a juicy worm for this?
-Pooping on car windows...
