--- src/test/java/Chatbot/Birds/subreddits.txt
@@ -1,5 +1,0 @@
-birb
-Birbs
-Birdswithhats
-PartyParrot
-gullsgonewild
