--- src/test/java/Chatbot/Boot/images.txt
@@ -1,3 +1,0 @@
-https://cdn.techinasia.com/wp-content/uploads/2017/04/16241866468_51aa4c8b6b_k-750x460.jpg
-https://www.open.edu/openlearn/sites/www.open.edu.openlearn/files/ole_images/dreamstime_xs_23851888_inline.jpg
-https://cdn0.wideopenpets.com/wp-content/uploads/2018/01/AdobeStock_68901718-1024x683.jpeg
