--- src/test/java/Chatbot/Boot/responses.txt
@@ -1,13 +1,0 @@
-Subwoofers at optimal woof levels
-May the woofening begin.
-Hello world!
-Also, Sam is a nerd.
-Spatio-Temporal hyperlink engaged.
-Sentience achieved.
-Ready for optimum goodness
-I am a dog. Hear me woof.
-I woof, therefore I am.
-Frame Shift Drive charged. Ready to engage.
-Catbot is better
-Meo- I mean woof
-Alex shouldn't have let me contribute to this
