--- src/test/java/Chatbot/Cats/responses.txt
@@ -1,5 +1,0 @@
-Pushing glasses off funiture...
-I'm only doing this if I get a treat
-Mewwwwww
-When's second breakfast?
-I better be getting a fishy for this
