--- src/test/java/Chatbot/Cats/subreddits.txt
@@ -1,5 +1,0 @@
-cats
-catspotting
-kittens
-kitties
-floof
