--- src/test/java/Chatbot/Dogs/extraGoodDogs.txt
@@ -1,100 +1,0 @@
-https://i.imgur.com/owJKr7y.jpg
-https://i.imgur.com/rpQdRoY.jpg
-https://i.imgur.com/Bt6zrhq.jpg
-https://i.imgur.com/sNm8hRR.jpg
-https://i.imgur.com/zMs8Sx6.jpg
-https://i.imgur.com/y48DeBZ.jpg
-https://i.imgur.com/6BCFIfW.jpg
-https://i.imgur.com/FBc1qrm.jpg
-https://i.imgur.com/l2Qx9Ba.jpg
-https://i.imgur.com/0Q8E82b.jpg
-https://i.imgur.com/QiudnCT.jpg
-https://i.imgur.com/uEHtSpB.jpg
-https://i.imgur.com/ZkxnaRE.jpg
-https://i.imgur.com/X8FO7Ds.jpg
-https://i.imgur.com/yKAXiyl.jpg
-https://i.imgur.com/CkJGGnc.jpg
-https://i.imgur.com/A4eVQQF.jpg
-https://i.imgur.com/Wtjxxiv.jpg
-https://i.imgur.com/uIX5RVf.jpg
-https://i.imgur.com/OYBBx0H.png
-https://i.imgur.com/t76aUWa.jpg
-https://i.imgur.com/VTEgzXg.jpg
-https://i.imgur.com/rFmyIDA.jpg
-https://i.imgur.com/mkeweJi.jpg
-https://i.imgur.com/YX3yu7h.jpg
-https://i.imgur.com/hhKzSwh.jpg
-https://i.imgur.com/jmu6xqa.jpg
-https://i.imgur.com/4WkDu3c.jpg
-https://i.imgur.com/BAUrawT.jpg
-https://i.imgur.com/pm50jbF.jpg
-https://i.imgur.com/SXKgmmG.jpg
-https://i.imgur.com/TmNmIf1.jpg
-https://i.imgur.com/knzWVG2.jpg
-https://i.imgur.com/UzvzMPd.jpg
-https://i.imgur.com/DDgeKjb.jpg
-https://i.imgur.com/12CuEvm.jpg
-https://i.imgur.com/vB2iE0u.jpg
-https://i.imgur.com/J7tUiy1.jpg
-https://i.imgur.com/vaDzijJ.jpg
-https://i.imgur.com/4KIZLwT.jpg
-https://i.imgur.com/1EocQiW.jpg
-https://i.imgur.com/gz7BCa1.jpg
-https://i.imgur.com/JbyflZV.jpg
-https://i.imgur.com/vBJ73wV.jpg
-https://i.imgur.com/Azrc9b6.jpg
-https://i.imgur.com/ZfIE7dc.jpg
-https://i.imgur.com/vxCSJ0n.jpg
-https://i.imgur.com/Xs4bOBt.jpg
-https://i.imgur.com/ller9Ft.jpg
-https://i.imgur.com/IoUuPka.jpg
-https://i.imgur.com/9YT06IC.jpg
-https://i.imgur.com/WYmvMnX.jpg
-https://i.imgur.com/Ts5NRT7.jpg
-https://i.imgur.com/Vq6ng5r.jpg
-https://i.imgur.com/en9mMT4.jpg
-https://i.imgur.com/8OdhqAf.jpg
-https://i.imgur.com/8rxgB9P.jpg
-https://i.imgur.com/jqaOwya.jpg
-https://i.imgur.com/lcuF8R5.jpg
-https://i.imgur.com/JkVzRrd.jpg
-https://i.imgur.com/aU8m1rK.jpg
-https://i.imgur.com/WpcyD30.jpg
-https://i.imgur.com/MwBVgf7.jpg
-https://i.imgur.com/dvbHqua.jpg
-https://i.imgur.com/ijTuUrr.jpg
-https://i.imgur.com/v0mgj4x.jpg
-https://i.imgur.com/ZToJLbs.jpg
-https://i.imgur.com/8bNYgmi.jpg
-https://i.imgur.com/jmu6xqa.jpg
-https://i.imgur.com/Wm4Ilnd.jpg
-https://i.imgur.com/rRGq4ke.jpg
-https://i.imgur.com/GUp6syL.jpg
-https://i.imgur.com/bnAioSK.jpg
-https://i.imgur.com/7loNOps.jpg
-https://i.redd.it/hohjvnmus7oz.jpg
-https://i.imgur.com/PaWEM3M.png
-https://i.imgur.com/J8RbiDc.jpg
-https://i.imgur.com/dEVaIWs.png
-https://i.imgur.com/asLgIfn.png
-https://i.imgur.com/e0nFdO9.jpg
-https://i.imgur.com/2FvMATu.jpg
-https://i.imgur.com/vaiFFmr.jpg
-https://i.imgur.com/kBdPrQd.png
-https://i.imgur.com/Y4ApsJh.jpg
-https://i.imgur.com/m83juDs.jpg
-https://i.imgur.com/FaEvM6d.jpg
-https://i.imgur.com/8A3EC1E.jpg
-https://i.imgur.com/xc6vt2k.png
-https://i.imgur.com/UxeWFns.jpg
-https://i.imgur.com/8uTiP82.jpg
-https://i.imgur.com/3RAv7JG.png
-https://i.imgur.com/CayjuKa.jpg
-https://i.imgur.com/OZthAcZ.png
-https://i.imgur.com/ZVo37II.png
-https://i.imgur.com/2NmaBlR.png
-https://i.imgur.com/12N8T7v.png
-https://i.imgur.com/J4vCKDk.png
-https://i.imgur.com/zOPBoU4.png
-https://i.imgur.com/4kPP2rj.png
-https://i.imgur.com/IcPQS31.jpg
