--- src/test/java/Chatbot/Dogs/responses.txt
@@ -1,16 +1,0 @@
-Browsing the pupternet...
-Loading dogs...
-Buffering optimum dogs..
-Reticulating Splines...
-Picking a very good dog..
-Doing the impossible...
-Chasing squirrels...
-Doin a heckin search...
-Optimising dogs...
-They always ask who's a good boy. But never how is the good boy.
-Loading handsome boy
-Loading gorgeous girle
-In space, no-one can hear you woof
-If you're reading this, it might be too late. Quick, they're comin-
-Wüüf
-
