--- src/test/java/Chatbot/Dogs/subreddits.txt
@@ -1,6 +1,0 @@
-rarepuppers
-dogswithjobs
-dog_irl
-woofbarkwoof
-wholesomedogmemes
-dogpictures
