--- src/test/java/Chatbot/Reacts/catReacts.txt
@@ -1,15 +1,0 @@
-https://hollandjake.com/dogbot/reacts/angry.jpg
-https://hollandjake.com/dogbot/reacts/confident.jpg
-https://hollandjake.com/dogbot/reacts/confused.jpg
-https://hollandjake.com/dogbot/reacts/disgusted.jpg
-https://hollandjake.com/dogbot/reacts/ecstatic.jpg
-https://hollandjake.com/dogbot/reacts/embarrassed.jpg
-https://hollandjake.com/dogbot/reacts/exhausted.jpg
-https://hollandjake.com/dogbot/reacts/frightened.jpg
-https://hollandjake.com/dogbot/reacts/frustrated.jpg
-https://hollandjake.com/dogbot/reacts/guilty.jpg
-https://hollandjake.com/dogbot/reacts/happy.jpg
-https://hollandjake.com/dogbot/reacts/hysterical.jpg
-https://hollandjake.com/dogbot/reacts/mischievous.jpg
-https://hollandjake.com/dogbot/reacts/sad.jpg
-https://hollandjake.com/dogbot/reacts/suspicious.jpg
